move util belt logic into a utils.belts module
